Fix issue with encoder Fix issue with PID controller

+++ lib/hardware/hw_encoder.c
@@ -26,7 +26,7 @@ void initialize_encoder(encoder_index_t index) {
 	}
 }
 
-unsigned int get_encoder_value(encoder_index_t index){
+uint16_t get_encoder_value(encoder_index_t index){
 
 	TIM_HandleTypeDef *htim = get_timer_handeler(encoder_info[index].timer);
 	return htim->Instance->CNT;
